Refactor: move `hash_script` to utility module
